# cmd/kubectl-audit-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-cluster aggregation server for kubectl-audit: clusters push findings.json scans here, which are deduplicated and stored for cross-cluster triage, centralized knowledge base/exclusion rules, and automation rule evaluation.",
        "title": "kubectl-audit-server API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/exclusion-rules": {
            "get": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Lists exclusion rules, optionally narrowed to one cluster (rules with no clusterId apply to every cluster).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exclusion-rules"
                ],
                "summary": "List exclusion rules",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Restrict to rules for this cluster ID",
                        "name": "cluster_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/internal_server_http.exclusionRuleDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid cluster_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid admin token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Creates a rule that suppresses matching findings, org-wide or for one cluster (clusterId).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exclusion-rules"
                ],
                "summary": "Create an exclusion rule",
                "parameters": [
                    {
                        "description": "Exclusion rule to create",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server_http.exclusionRuleDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_server_http.exclusionRuleDTO"
                        }
                    },
                    "400": {
                        "description": "invalid request body / missing reason / invalid clusterId",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid admin token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/exclusion-rules/{id}": {
            "delete": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Deletes one exclusion rule by ID.",
                "tags": [
                    "exclusion-rules"
                ],
                "summary": "Delete an exclusion rule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Exclusion rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "deleted"
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid admin token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "no such exclusion rule",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ingest/native": {
            "post": {
                "security": [
                    {
                        "ClusterAuth": []
                    }
                ],
                "description": "Ingests a kubectl-audit native findings.json body for the cluster identified by the bearer token. May produce more than one Scan if the payload batches multiple sources.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingest"
                ],
                "summary": "Ingest a native findings.json scan",
                "parameters": [
                    {
                        "description": "findings.json payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server_http.ingestResponse"
                        }
                    },
                    "400": {
                        "description": "invalid payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid bearer token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ingest/openreports": {
            "post": {
                "security": [
                    {
                        "ClusterAuth": []
                    }
                ],
                "description": "Ingests an openreports.io/v1alpha1 Report or ClusterReport JSON body for the cluster identified by the bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingest"
                ],
                "summary": "Ingest an OpenReports scan",
                "parameters": [
                    {
                        "description": "OpenReports Report/ClusterReport payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_server_http.ingestResponse"
                        }
                    },
                    "400": {
                        "description": "invalid payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid bearer token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_server_http.exclusionMatchDTO": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
        "internal_server_http.exclusionRuleDTO": {
            "type": "object",
            "properties": {
                "clusterId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "match": {
                    "$ref": "#/definitions/internal_server_http.exclusionMatchDTO"
                },
                "policyIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "internal_server_http.ingestResponse": {
            "type": "object",
            "properties": {
                "scans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_server_http.scanResult"
                    }
                }
            }
        },
...
        "internal_server_http.scanResult": {
            "type": "object",
            "properties": {
                "findingsIngested": {
                    "type": "integer"
                },
                "scanId": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin bearer token (env ADMIN_TOKEN). Send as \"Bearer \u003ctoken\u003e\". Required for cluster registration and every organization-level config endpoint (knowledge base, exclusion rules, automation rules, triage).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ClusterAuth": {
            "description": "Per-cluster bearer token minted by POST /clusters. Send as \"Bearer \u003ctoken\u003e\". Required for ingest endpoints only — scoped to that cluster's own data, and does not grant triage read/write (see AdminAuth).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}